refactor(output): 당첨 통계 출력 기능 outputService.js로 이동
outputService 파일에서 가지고 있는 것이 적합하다고 판단
